% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_proxy.R
\name{brain_proxy}
\alias{brain_proxy}
\title{Shiny Proxy for Viewer}
\usage{
brain_proxy(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{shiny output ID}

\item{session}{shiny session, default is current session (see \code{\link[shiny]{getDefaultReactiveDomain}})}
}
\value{
\code{R6} class \code{ViewerProxy}
}
\description{
Shiny Proxy for Viewer
}
