% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{brain_setup}
\alias{brain_setup}
\title{Setup Package, Install Environment}
\usage{
brain_setup(
  continued = FALSE,
  show_example = TRUE,
  use_python = FALSE,
  try_conda = TRUE
)
}
\arguments{
\item{continued}{logical, there are two phases of setting up environment. You
probably need to restart R session after the first phase and continue setting up.}

\item{show_example}{whether to show example of `N27` subject at the end.}

\item{use_python}{whether to install python toolbox (recommended, but not by default)}

\item{try_conda}{try to use `conda` to create `RAVEPy` environment}
}
\description{
Setup Package, Install Environment
}
\author{
Zhengjia Wang
}
