% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_orient.R
\name{calculate_rotation}
\alias{calculate_rotation}
\title{Calculate rotation matrix from non-zero vectors}
\usage{
calculate_rotation(vec_from, vec_to)
}
\arguments{
\item{vec_from}{original vector, length of 3}

\item{vec_to}{vector after rotation, length of 3}
}
\value{
A four-by-four transform matrix
}
\description{
Calculate rotation matrix from non-zero vectors
}
