% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_orient.R
\name{conform_volume}
\alias{conform_volume}
\title{Conform imaging data in \code{'FreeSurfer'} way}
\usage{
conform_volume(x, save_to, dim = c(256, 256, 256))
}
\arguments{
\item{x}{path to the image file}

\item{save_to}{path where the conformed image will be saved, must ends with
\code{'.mgz'}}

\item{dim}{positive integers of length three, the conformed dimension;
by default 'FreeSurfer' conform images to \code{1mm} volume cube with
\code{256x256x256} dimension}
}
\value{
Nothing; the result will be save to \code{save_to}
}
\description{
Reproduces conform algorithm used by \code{'FreeSurfer'} to conform
\code{'NIfTI'} and \code{'MGH'} images.
}
