% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_orient.R
\name{cross_prod}
\alias{cross_prod}
\title{Calculate cross-product of two vectors in '3D'}
\usage{
cross_prod(x, y)
}
\arguments{
\item{x, y}{3-dimensional vectors}
}
\value{
A '3D' vector that is the cross-product of \code{x} and \code{y}
}
\description{
Calculate cross-product of two vectors in '3D'
}
