% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_template_directory}
\alias{default_template_directory}
\title{Default Directory to Store Template Brain}
\usage{
default_template_directory(check = FALSE)
}
\arguments{
\item{check}{logical, check if the folder is missing, is so,
create one. This option ensures the folder is always created.}
}
\value{
A directory path where template brain is stored at; see also
\code{\link{download_N27}}
}
\description{
Default Directory to Store Template Brain
}
\details{
When \code{threeBrain.template_dir} is not set or invalid, the
function checks 'RAVE' (R Analysis and Visualization for 'iEEG',
\url{https://openwetware.org/wiki/RAVE}) folder at home directory. If
this folder is missing, then returns results from
\code{R_user_dir('threeBrain', 'data')}. To override the default behavior,
use \code{options(threeBrain.template_dir=...)}.
}
\examples{

default_template_directory()

}
