% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rave.R
\name{dev_threeBrain}
\alias{dev_threeBrain}
\title{Function to load all dev funtions and wrap them within an environment}
\usage{
dev_threeBrain(expose_functions = FALSE, reload = TRUE)
}
\arguments{
\item{expose_functions}{logical indicating whether to expose all dev functions to the global environment}

\item{reload}{logical, do you want to fast-reload the package before load the functions?}
}
\description{
Function to load all dev funtions and wrap them within an environment
}
