% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{download_N27_surface}
\alias{download_N27_surface}
\title{Download N27 Brain Surfaces From URL}
\usage{

  download_N27_surface(url_root = "https://s3.amazonaws.com/rave-data/sample-data/N27/",
  surfaces = c("pial", "white", "sphere", "smoothwm", "inflated"),
  reset = FALSE, n27_dir = getOption("threeBrain.n27_dir",
  "~/rave_data/others/three_brain/N27"))
}
\arguments{
\item{url_root}{URL address to store the meshes}

\item{surfaces}{which surface(s) to download}

\item{reset}{reset downloads?}

\item{n27_dir}{download to}
}
\description{
Download N27 Brain Surfaces From URL
}
