% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_outer_smooth.R
\name{fill_surface}
\alias{fill_surface}
\title{Fill in a surface with given volume size}
\usage{
fill_surface(
  surface,
  save_as = NULL,
  resolution = 256,
  delta = 2,
  IJK2RAS = NULL,
  verbose = TRUE
)
}
\arguments{
\item{surface}{a 'FreeSurfer' surface object, or file path to the object,
or a \code{'rgl'} \code{'mesh3d'} object; must be closed, water-tight}

\item{save_as}{path to file where the results should be saved; the format
is 'FreeSurfer' \code{'mgz'}; default is \code{NULL} (no save)}

\item{resolution}{volume resolution; default is to store results in a
\code{256 x 256 x 256} cube}

\item{delta}{threshold to remove isolated points that are wrongfully created}

\item{IJK2RAS}{a \code{4 x 4} volume 'IJK' index to anatomical 'RAS'
transform; default is automatically determined}

\item{verbose}{whether to verbose the progress; default is true}
}
\value{
A list of filled volume with \code{IJK2RAS} matrix; if
\code{save_as} is valid, a \code{'mgz'} file will be saved
}
\description{
Ray-marching through each margin to find a volume that are
inside of a water-tight mesh
}
\examples{


library(threeBrain)
if(file.exists(file.path(default_template_directory(), "N27"))) {

  lh_pial <- file.path(default_template_directory(),
                       "N27", "surf", "lh.pial.asc")
  save_as <- tempfile(fileext = ".mgz")

  fill_surface(lh_pial, save_as)

  # you can view `save_as` in FreeSurfer viewer

}


}
