% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_lut.R
\docType{data}
\name{freesurfer_lut}
\alias{freesurfer_lut}
\title{Query the 'FreeSurfer' labels}
\format{
An object of class \code{list} of length 2.
}
\usage{
freesurfer_lut
}
\description{
Query the 'FreeSurfer' labels
}
\details{
The 'FreeSurfer' atlases use
\url{https://surfer.nmr.mgh.harvard.edu/fswiki/FsTutorial/AnatomicalROI/FreeSurferColorLUT}
look-up table to query indexes. The 'threeBrain' electrode localization
also uses this table to export electrode 'FSLabel'. If volume type is
set to \code{'aparc_aseg'}, then please also use this table to filter.
}
\examples{

freesurfer_lut$from_key(0:10)

freesurfer_lut$get_key("ctx-lh-supramarginal")

}
\keyword{datasets}
