% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_outer_smooth.R
\name{generate_smooth_envelope}
\alias{generate_smooth_envelope}
\title{Generate smooth envelope around surface}
\usage{
generate_smooth_envelope(
  filled_volume_path,
  save_as,
  inflate = 2,
  verbose = TRUE
)
}
\arguments{
\item{filled_volume_path}{path to \code{'mgz'} with filled volume, see
\code{save_as} in \code{\link{fill_surface}}}

\item{save_as}{save final envelope to path}

\item{inflate}{number of \code{'voxels'} to inflate before fitting envelope;
must be a non-negative integer}

\item{verbose}{whether to verbose the progress; default is true}
}
\value{
Nothing, the result is saved to \code{save_as}
}
\description{
Alternative to 'Matlab' version of \code{'pial-outer-smoothed'},
use this function along with \code{\link{fill_surface}}.
}
\examples{

library(threeBrain)
if(file.exists(file.path(default_template_directory(), "N27"))) {

  lh_pial <- file.path(default_template_directory(),
                       "N27", "surf", "lh.pial.asc")
  tmp_volume <- tempfile(fileext = ".mgz")

  fill_surface(lh_pial, tmp_volume)

  save_as <- tempfile(fileext = ".pial-outer-smoothed")
  generate_smooth_envelope(tmp_volume, save_as)

  # see file `save_as` as leptomeningeal approximation

}

}
