% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generate_subcortical_surf.R
\name{get_ijk2ras}
\alias{get_ijk2ras}
\title{Get 'voxel' to world matrix}
\usage{
get_ijk2ras(x, type = c("scanner", "tkr"))
}
\arguments{
\item{x}{path to imaging files}

\item{type}{world space type; choices are \code{'scanner'} (same as
\code{'sform'} or \code{'qform'} in) or \code{'NIfTI'} file headers;
or \code{'tkr'} (used to shared surface nodes)}
}
\value{
A four by four matrix
}
\description{
Get 'voxel' to world matrix
}
