% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_subject.R
\name{import_from_freesurfer}
\alias{import_from_freesurfer}
\title{Import from `FreeSurfer` and create `JSON` cache for 3D viewer}
\usage{
import_from_freesurfer(fs_path, subject_name)
}
\arguments{
\item{fs_path}{`FreeSurfer` subject directory}

\item{subject_name}{subject code}
}
\value{
None.
}
\description{
Import from `FreeSurfer` and create `JSON` cache for 3D viewer
}
