% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_electrode_shaft.R
\name{new_electrode_prototype}
\alias{new_electrode_prototype}
\title{Create a prototype for visualizing electrode shaft}
\usage{
new_electrode_prototype(
  prefix = "",
  min_length = 0,
  geom_control_points = NULL,
  geom_control_diameters = NULL,
  contacts = list()
)
}
\arguments{
\item{prefix}{geometry prefix}

\item{min_length}{minimum length in millimeters}

\item{geom_control_points}{control points in millimeters, used to control the
shape of the electrode, with zero being the target point.}

\item{geom_control_diameters}{diameter for each control point in millimeters,
must share the same length as \code{geom_control_points}}

\item{contacts}{a list of electrode positions; see example}
}
\description{
Create a prototype for visualizing electrode shaft
}
\examples{

proto <- new_electrode_prototype(
  prefix = "sDEA-2mm-",
  geom_control_points = c(0, 1.2, 10.2, 10.201, -3, -2, -0.1),
  geom_control_diameters = c(0, 0.2, 0.2, 1.1, 1.1, 1.5, 1.5),
  contacts = list(
    list(
      order = 1,
      start_position = 0.2,
      size = 10,
      description = "microwire"
    ),
    list(
      order = 2,
      start_position = 10.2,
      size = 1.32,
      description = "macro-1"
    ),
    list(
      order = 3,
      start_position = 12.2,
      size = 1.32,
      description = "macro-2"
    ),
    list(
      order = 4,
      start_position = 14.2,
      size = 1.32,
      description = "macro-3"
    ),
    list(
      order = 5,
      start_position = 16.2,
      size = 1.32,
      description = "macro-4"
    )
  )
)

geom <- proto$generate_geometry(
  "shaft-1",
  subject_code = "N27",
  target_position = c(8.6, 38.7, -42.8),
  entry_position = c(49, 49.3, -25.5),
  length = 100
)

n27_path <- file.path(
  default_template_directory(),
  "N27"
)
if(interactive() && file.exists(n27_path)) {
  brain <- threeBrain(n27_path, subject_code = "N27")
  brain$plot(additional_geoms = list(geom))
}

}
