% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_fs_asc}
\alias{read_fs_asc}
\title{Read `FreeSurfer` ascii file}
\usage{
read_fs_asc(file)
}
\arguments{
\item{file}{file location}
}
\value{
a list of vertices and face indices
}
\description{
Read `FreeSurfer` ascii file
}
