% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_fs_labels}
\alias{read_fs_labels}
\title{Read FreeSurfer Annotations}
\usage{
read_fs_labels(path, vertex_number)
}
\arguments{
\item{path}{label path}

\item{vertex_number}{force to reset vertex number if raw file is incorrect}
}
\description{
Read FreeSurfer Annotations
}
