% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_fs_mgh_mgz}
\alias{read_fs_mgh_mgz}
\title{Read `FreeSurfer` `mgz/mgh` file}
\usage{
read_fs_mgh_mgz(filename)
}
\arguments{
\item{filename}{file location}
}
\value{
list contains coordinate transforms and volume data
}
\description{
Read `FreeSurfer` `mgz/mgh` file
}
