% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{read_gii2}
\alias{read_gii2}
\title{Function to load surface data from `Gifti` files}
\format{
An \R function acting as safe wrapper for \code{nibabel.load}.
}
\usage{
read_gii2(path)
}
\arguments{
\item{path}{`Gifti` file path}
}
\description{
The function `read_gii2` is a dynamic wrapper of Python `nibabel` loader.
If no Python is detected, it will switch to `gifti::readgii`.
}
