% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{read_mgz}
\alias{read_mgz}
\title{Function to load `FreeSurfer` `mgz/mgh` file}
\format{
An \R function acting as safe wrapper for \code{nibabel.load}.
}
\usage{
read_mgz(path)
}
\arguments{
\item{path}{`mgz/mgh` file path}
}
\description{
The function `read_mgz` is a dynamic wrapper of Python `nibabel` loader.
If no Python is detected, it will switch to built-in function `read_fs_mgh_mgz`,
which has limited features.
}
