% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generate_subcortical_surf.R
\name{read_volume}
\alias{read_volume}
\title{Read volume file in \code{'MGH'} or \code{'Nifti'} formats}
\usage{
read_volume(file, format = c("auto", "mgh", "nii"), header_only = FALSE)
}
\arguments{
\item{file}{file path}

\item{format}{the file format}

\item{header_only}{whether only read headers; default is false}
}
\value{
A list of volume data and transform matrices; if
\code{header_only=TRUE}, then volume data will be substituted by the
header.
}
\description{
Read volume file in \code{'MGH'} or \code{'Nifti'} formats
}
