% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{renderBrain}
\alias{renderBrain}
\title{Shiny Renderer for threeBrain Widgets}
\arguments{
\item{expr}{R expression that calls three_brain function or Brain object}

\item{env}{environment of expression to be evaluated}

\item{quoted}{is expr quoted? Default is false.}
}
\description{
Shiny Renderer for threeBrain Widgets
}
