% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{save_brain}
\alias{save_brain}
\title{Save threeBrain widgets to local file system}
\usage{
save_brain(
  widget,
  directory,
  filename = "index.html",
  assetpath = "lib/",
  datapath = "lib/threebrain_data-0/",
  title = "3D Viewer",
  as_zip = FALSE
)
}
\arguments{
\item{widget}{generated from function 'threejs_brain'.}

\item{directory}{directory to save the widget.}

\item{filename}{default is 'index.html', filename of the widget index file.}

\item{assetpath}{where to put \code{css} or \code{JavaScript} to,
must be relative to \code{directory}.}

\item{datapath}{where to store data to, must be relative to \code{directory}.}

\item{title}{widget title.}

\item{as_zip}{whether to create zip file "compressed.zip".}
}
\description{
Save threeBrain widgets to local file system
}
\author{
Zhengjia Wang
}
