% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer_ctint1.R
\name{view_ct_t1}
\alias{view_ct_t1}
\title{View CT with T1 image}
\usage{
view_ct_t1(
  subject_code,
  fs_path,
  ct_path = file.path(fs_path, "RAVE", "coregistration", "ct2t1.nii.gz")
)
}
\arguments{
\item{subject_code}{subject code}

\item{fs_path}{FreeSurfer subject path}

\item{ct_path}{where CT file is stored, require 'Nifti' format}
}
\description{
View aligned CT scan with T1 images
}
