% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_blank.R
\docType{data}
\name{BlankGeom}
\alias{BlankGeom}
\title{A geomery that renders nothing}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BlankGeom
}
\description{
This is mainly used when you want to upload group data only
}
\keyword{datasets}
