% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_datacube2.R
\name{DataCubeGeom2}
\alias{DataCubeGeom2}
\title{R6 Class - Generate Data Cube Geometry via 3D Volume Texture}
\description{
R6 Class - Generate Data Cube Geometry via 3D Volume Texture
}
\author{
Zhengjia Wang
}
