% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_brain.R
\name{freesurfer_brain}
\alias{freesurfer_brain}
\title{Read `FreeSurfer` surface and volume files}
\usage{
freesurfer_brain(fs_subject_folder, subject_name,
  additional_surfaces = NULL, use_cache = TRUE, use_141 = TRUE)
}
\arguments{
\item{fs_subject_folder}{character, `FreeSurfer` subject folder, or `RAVE` subject folder}

\item{subject_name}{character, subject code to display with only letters and digits}

\item{additional_surfaces}{character array, additional surface types to load, such as `white`, `smoothwm`}

\item{use_cache}{logical, whether to use cached json files or from raw `FreeSurfer` files}

\item{use_141}{logical, whether to use standard 141 brain for surface file}
}
\description{
Read `FreeSurfer` surface and volume files
}
\details{
This function is under FreeSurfer license.
1. Volumes:
3D viewer uses `mri/brain.finalsurfs.mgz` from `FreeSurfer` to show the
volume information. `brain.finalsurfs.mgz` results from step 1 to 15 in `FreeSurfer`
command `recon-all`, which aligns the original `DICOM` image to `RAS` coordinate
system, resamples to volume with \code{256x256x256} voxels (trilinear by default,
check \url{https://mail.nmr.mgh.harvard.edu/pipermail/freesurfer/2018-August/058127.html}
for more information).

2. Surface:
There are two options for surface files. The first choice is using
`std.141` brain generated by `AFNI/SUMA`. This surface file re-calculates vertices
from standard 141 space, which averages the "surface" of 141 subjects. If you
want to map surface electrodes across different subjects, you might want to
consider this case as it's especially designed for surface mapping. However,
you'll need `AFNI/SUMA` installed to generate the surface file. The details
can be found via \url{https://openwetware.org/wiki/Beauchamp:CorticalSurfaceHCP},
and the `AFNI/SUMA` command related is `SurfToSurf`. Please generate the files
to `[FREESURFER SUBJECT DIR]/SUMA/`. The file name follows the convention of
`std.141.[lr]h.[SURFACE TYPE].[POSTFIX]`, where `lh` means left hemisphere and
`rh` means right hemisphere; `SURFACE TYPE` can be `pial`, `white`, `smoothwm`,
and `POSTFIX` can be `asc`, `gii`. If multiple files for the same surface
type exists, the search order will be `asc > gii`.
The other option is to use mesh files directly from freesurfer output located
at `[FREESURFER SUBJECT DIR]/surf`. If you want to use these surface, make sure
they are converted to `asc` or `gii` format.

3. Electrode registration and transforms
This package provides two ways to map electrodes to standard space. For surface
electrodes, if standard 141 brain is provided, then the first option is to snap
electrodes to the nearest vertices in subject space. The key is the vertex number
matches across different subjects, hence the location of corresponding vertices
at template brain are the mapped electrode coordinates.
If standard 141 brain is missing, or the electrode type is `stereo EEG`, then
the second option is volume mapping. The idea is to map electrodes to `MNI305`
brain. The details can be found at \url{https://surfer.nmr.mgh.harvard.edu/fswiki/CoordinateSystems}.
To perform volume mapping, we need `FreeSurfer` folder `mri/tranforms`.
Currently, only linear `Talairach` transform matrix is supported (located at
`talairach.xfm`).

4. Coordinates
The 3D viewer in this package uses the center of volume as the origin (0, 0, 0).
}
\author{
Zhengjia Wang
}
