% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_electrode_proto.R
\name{list_electrode_prototypes}
\alias{list_electrode_prototypes}
\title{List all electrode prototypes}
\usage{
list_electrode_prototypes()
}
\value{
A named list, names are the prototype types and values are the
prototype configuration paths.
}
\description{
List all built-in and user-customized electrode prototypes. User paths
will be searched first, if multiple prototype configuration files are found
for the same type.
}
\examples{

list_electrode_prototypes()


}
