% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_multibrain.R
\name{merge_brain}
\alias{merge_brain}
\title{Create Multi-subject Template}
\usage{
merge_brain(
  ...,
  .list = NULL,
  template_surface_types = NULL,
  template_subject = unname(getOption("threeBrain.template_subject", "N27")),
  template_dir = unname(getOption("threeBrain.template_dir",
    "~/rave_data/others/three_brain"))
)
}
\arguments{
\item{..., .list}{\code{Brain2} objects}

\item{template_surface_types}{which template surface types to load, default is auto-guess}

\item{template_subject}{character, subject code to be treated as template, default is `N27`}

\item{template_dir}{the parent directory where template subject is stored in}
}
\description{
Create Multi-subject Template
}
\author{
Zhengjia Wang
}
