% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{save_brain}
\alias{save_brain}
\title{Save threeBrain widgets to local file system}
\usage{
save_brain(widget, path, title = "3D Viewer", as_zip = FALSE, ...)
}
\arguments{
\item{widget}{generated from function 'threejs_brain'}

\item{path}{path to save the brain widget}

\item{title}{widget title.}

\item{as_zip}{whether to create zip file "compressed.zip".}

\item{...}{ignored, used for backward compatibility}
}
\description{
Save threeBrain widgets to local file system
}
\author{
Zhengjia Wang
}
