% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{threejs_brain}
\alias{threejs_brain}
\title{Create a Threejs Brain}
\usage{
threejs_brain(..., time_range = NULL, value_range = NULL,
  symmetric = 0, side_camera = FALSE, control_panel = TRUE,
  control_presets = NULL, camera_center = c(0, 0, 30),
  color_ramp = c("navyblue", "#e2e2e2", "red"), n_color = 100,
  show_legend = TRUE, legend_title = "Value", legend_expr, at = NULL,
  width = NULL, height = NULL, .list = list())
}
\description{
Create a Threejs Brain
}
