% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{threejs_brain}
\alias{threejs_brain}
\title{Create a Threejs Brain}
\arguments{
\item{..., .list}{geometries inherit from AbstractGeom}

\item{widget_id}{unique identifier for the widget. Use it when you have
multiple widgets in one website (shiny for example)}

\item{time_range}{used to calculate animation time (not yet implemented)}

\item{value_range}{used to generate colors}

\item{symmetric}{default 0, color center will be mapped to this value}

\item{side_camera}{enable side cameras to view objects from fixed perspective}

\item{control_panel}{enable control panels for the widget}

\item{camera_center}{position where camera should focus at}

\item{color_ramp}{used to generate color ramps}

\item{n_color}{how many colors in the color ramp (used  when generating legend)}

\item{show_legend}{show legend in control panel?}

\item{legend_title}{legend title}

\item{at}{legen ticks}

\item{legend_expr}{R expression to generate legend if you don't like the
default ones}

\item{tmp_dirname}{internally used}

\item{width, height}{width and height of the widget. By default width="100%",
and height varies.}
}
\description{
Create a Threejs Brain
}
