% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejs_brain.R
\name{threejsBrainOutput}
\alias{threejsBrainOutput}
\title{Shiny Output for threeBrain Widgets}
\arguments{
\item{outputId}{unique identifier for the widget}

\item{width, height}{width and height of the widget. By default width="100%",
and height="500px".}

\item{reportSize}{whether to report widget size in shiny
\code{session$clientData}}
}
\description{
Shiny Output for threeBrain Widgets
}
\author{
Zhengjia Wang
}
