% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palette_voxel.R
\name{voxel_palette}
\alias{voxel_palette}
\alias{freeserfer_palette}
\alias{create_voxel_palette_discrete}
\alias{create_voxel_palette_continuous}
\title{Creates color palettes for volume data (\code{\link{DataCubeGeom2}})}
\usage{
freeserfer_palette(write_to = NULL)

create_voxel_palette_discrete(
  key,
  color,
  label = NULL,
  alpha = FALSE,
  con = NULL,
  ...
)

create_voxel_palette_continuous(
  key,
  color,
  value = NULL,
  alpha = FALSE,
  con = NULL,
  ...
)
}
\arguments{
\item{key}{integer vector corresponding to voxel values. If the actual
values are not integer, please also assign \code{value}}

\item{color}{characters, corresponding to color strings for each key}

\item{label}{name for each key, discrete palettes only}

\item{alpha}{whether to respect transparency}

\item{con, write_to}{a file path to write results to. The file path can be
passed as \code{voxel_palette} into \code{\link{threejs_brain}}.}

\item{...}{ignored}

\item{value}{actual value for each key, continuous palettes only}
}
\value{
A list of palette information
}
\description{
Creates color palettes for volume data (\code{\link{DataCubeGeom2}})
}
\examples{

# Creates a symmetric continuous palette with 3 keys
# The color range is -10 to 10
# The colors are 'blue','white','red' for these keys

create_voxel_palette_continuous(
  key = c(1,2,3), value = c(-10,0,10),
  color = c('blue','white','red'))

}
